Reorganize dialogs into folder
