+++ src/Practitioner/PatientProfile/Dialogs/ArchiveDialog.js
@@ -1,14 +1,14 @@
 import React from 'react'
 import { makeStyles } from '@material-ui/core/styles'
-import useStores from '../../Basics/UseStores'
-import PopOver from '../Shared/PopOver'
+import useStores from '../../../Basics/UseStores'
+import PopOver from '../../Shared/PopOver'
 import Typography from '@material-ui/core/Typography'
 import { useTranslation } from 'react-i18next'
-import DatePicker from '../../Basics/DatePicker'
+import DatePicker from '../../../Basics/DatePicker'
 import { observer } from 'mobx-react'
-import ProfileButton from '../../Components/FlatButton'
+import ProfileButton from '../../../Components/FlatButton'
 import CheckIcon from '@material-ui/icons/PlaylistAddCheck';
-import Colors from '../../Basics/Colors'
+import Colors from '../../../Basics/Colors'
 import TreatmentOutcomeSelection from './TreatmentOutcomeSelection'
 import {useLocation, useHistory} from 'react-router-dom';
