+++ src/Practitioner/PatientProfile/index.js
@@ -11,8 +11,8 @@ import SymptomSummary from './SymptomSummary';
 import TreatmentTimeline from '../../Basics/TreatmentTimeline';
 import ReportingHistory from './ReportingHistory';
 import SectionLabel from '../../Components/SectionLabel';
-import PatientProfileDialogs from './Dialogs';
-import ArchivedOptions from './ArchivedOptions';
+import PatientProfileDialogs from './Dialogs/';
+import ArchivedOptions from './Dialogs/ArchivedOptions';
 import ProfileHeader from './Header';
 import useWindowSize from '../../Hooks/Resize';
 import MobileView from './MobileView';
